+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-06-18 19:34:34
-// Commit: chore(ui): update dependencies and cleanup unused imports in ui
+// Updated at: 2026-06-18 22:14:49
+// Commit: docs(ui): update readme and API contracts for ui
 
 class uiComponentState {
-  final String commitRef = '2107';
-  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
-  final DateTime timestamp = DateTime.parse('2026-06-18T19:34:34+05:00');
+  final String commitRef = '2110';
+  final String lastMessage = 'docs(ui): update readme and API contracts for ui';
+  final DateTime timestamp = DateTime.parse('2026-06-18T22:14:49+05:00');
 }